fix: Update version of Nvidia driver in cloud images.
Upgrade from 410.104 to 450.51.05. This apparently is necessary to get
T4 chips to work on G4 AWS instances. This driver *should* be compatible
with older versions of CUDA -- I tested with CUDA 10.0 only.

Along the way, switch from the "XFree86" package to the "Tesla" package.
It is a little unclear what the difference is, but the Nvidia download
site guides you to the latter when you're using enterprise GPUs, which
should always be the case in public cloud environments.