Subject: Update AWS plugin build, NCCL and GDRcopy
Description:

This mod changes how NCCL is built to try updating to a more recent
version. It also tries to enable shared linking for libnccl to enable
users to override it with one they mount into the container.

This mod also adds libgdrapi* libraries to enable the AWS plugin to
leverage them. It also pulls the modified AWS plugin code to use.