build: add AWS OFI plugin for NCCL GPU based OFI/libfabric builds (#189)
* PAD-232 Initial AWS OFI Plugin

* Subject: Add OMPI+OFI support to GPU build and enable scraping host fabric

Description:

This mod enables the OMPI+OFI build for GPUs the same as is allowed
for CPUs. Further, it adds an entrypoint for both the tf2-cpu and
tf2-gpu images that can scrape up the host libfabric if
requested. This is intended to replace the libfabric in the container
to enable optimal performance on different architectures (non-IB).

The Horovod build on GPUs was also modified to use MPI for GPU
operations for now until NCCL is supported with libfabric.

Tested the OSU get bw on a libfabric based machine using both the GPU
and CPU images and verified that when scraping up the host libfabric
the performance matches expectations. For example:

user@nid000001:/scratch2/user> env USE_HOST_LIBFABRIC=y srun -n 2 --ntasks-per-node=1 --mpi=pmi2 singularity run --bind /scratch2,/usr:/host/usr /scratch2/user/sif/environments_py-3.8-pytorch-1.11-tf-2.8-cpu-mpi-ofi-0.19.4.sif ./osu-micro-benchmarks-5.6.2/mpi/one-sided/osu_get_bw
1                       0.94
2                       1.84
4                       3.80
8                       7.42
16                     15.22
32                     29.77
64                     61.05
128                   118.42
256                   233.74
512                   450.50
1024                  913.88
2048                 1788.54
4096                 3175.02
8192                 4167.06
16384                5227.06
32768                6122.97
65536               18987.43
131072              20382.48
262144              22252.80
524288              23231.07
1048576             23403.87
2097152             23698.02
4194304             23799.45

The important parts are the setting of the USE_HOST_LIBFABRIC=y to
tell the entry point to look for the host libfabric.so to pull into
the container, as well as mounting the location of the host libfabric
under /host inside the container (ie, /usr:/host/usr above).

* Add script for scraping host libfabric

* Add nccl tag, fix horovod build error

* Add AWS enable trace, set ld_lib_path

* Parameterize hardcoded paths

* Scrape libs if /host mounted, prepend aws to lib path

* Fix non OFI build

* PAD-232 Use NVidias code

* Update internal-aws-dataserver logic

* restrict changes to OFI/AWS builds

* Fix non-mpi gpu build failure

* Fix non-nccl gpu build failure

* Fix non-ofi gpu build path

* Do not pollute all images with unrelated INFO msgs

* Isolate entrypoint scrape-lib script

* Address review comments

---------

Co-authored-by: Chris Rickett <cdrickett@gmail.com>