fix: revert libnss_determined-goes-first
It appears that placing our libnss plugin first is good for openshift,
but bad for slurm.

If we have to pick, we should optimize for the slurm users.

Additionally, placing our plugin last offers the least surprise, I
think, as it doesn't override whatever else the user had configured in
their image (which was the original motivation for having it last).

Our install_libnss_determined.sh now accepts a --first option to get the
openshift-friendly behavior.  This is likely more useful as
documentation strategy than for any real users.